requirements.txt
streamlit>=1.32.0
pydantic>=2.5.0
numpy>=1.26.0
pandas>=2.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.5.0
requests>=2.31.0
scikit-learn>=1.3.0

# Optional: DROID metadata ingestion via HuggingFace (works on Python 3.14).
# Uncomment OR install on demand: `pip install datasets`
# datasets>=2.16.0

# Optional: TFDS DROID backend (TensorFlow ships wheels for Python 3.9-3.13 only).
# `pip install tensorflow tensorflow-datasets`
